requirements.txt
python-dotenv
openai
qdrant-client
PyMuPDF

